Complete EDF
Now EDF is exactly the same as the old one, with all the menus and all
that, only with small differences such as the name of the attachment
objects and the name of the extension.

All the EDF variables are still not initialized in the VB code, but
apparently it works normally.